Added EzySlice
The last implementation of mesh slicing didn't work very well for what I was trying to do, so I have decided to use EzySlice instead. I was hesitant at first to use it as there doesn't seem to be much documentation. but after finding a youtube video explaining how it can be implemented in VR, I now understand how to perform basic tasks such as mesh slicing, adding rigidbodies to sliced meshes and applying a material to the newly created faces.